% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortNormalizedExpression.R
\name{addExonExpression}
\alias{addExonExpression}
\title{Add exon expression file data to an existing data frame}
\usage{
addExonExpression(exonModels, id, path, type = "rpkm")
}
\arguments{
\item{exonModels}{The existing exon models data frame to append to, with or
without additional columns. The data.frame returned by this function is a
valid input for this parameter. See also \code{\link{loadCohortDefinition}}}

\item{id}{The name of the column in the data frame after it is appended}

\item{path}{The path to the exon expression data file being added.}

\item{type}{The name of the column in the exon expression data file that is
extracted and appended. By default this is \code{"rpkm"}}
}
\value{
A data frame with one row per exon with the initial columns
describing the exons (as loaded by \code{\link{loadCohortDefinition}}), any
following columns of data, and then the specified column from the given
exon expression data file appended.
}
\description{
Adds a column from a single exon data file to gene models data frame or an
exon expression data frame. Can be used to grow a cohort-wide file of
expression data one sample at a time by recursively using the returned data
frame as the input \code{exonModels} file.
}

