% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortFusions.R
\name{do.plots}
\alias{do.plots}
\title{Plot fusion gene expression}
\usage{
do.plots(fusionsDf, samples, geneOrderList, normalizedCohortExpressionDf)
}
\arguments{
\item{fusionsDf}{The fusions to plot}

\item{samples}{Vector of samples to plot}

\item{geneOrderList}{Vector of genes to plot, in priority order. Genes earlier in list
print on left side when paired with gene later in list.}

\item{normalizedCohortExpressionDf}{The cohort of normalized exon expression results.}
}
\value{
A named vector with three components:
   \tabular{ll}{
      \code{plots} \tab Plot count\cr
      \code{fusionPairs} \tab Number of paired genes plotted\cr
      \code{fusionSingles} \tab Number of single genes plotted\cr
   }
}
\description{
Plot fusion gene expression
}

