% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gafGeneModels.R
\name{extractGeneModels}
\alias{extractGeneModels}
\title{Extract gene models from the gaf}
\usage{
extractGeneModels(gafFile, outFile = paste0(gafFile, ".geneModels"),
  force = FALSE, uniqueGene = TRUE, skipUnknownGene = TRUE)
}
\arguments{
\item{gafFile}{The full-path name of the GAF file [REQ]}

\item{outFile}{The output file name, by default the same as the
gafFile, with ".geneModels" appended. Will not overwrite unless force= is
set true}

\item{force}{If the output file exists, overwrite (generates
a warning).}

\item{uniqueGene}{The gaf file contains several versions of
some genes, these are annotated as  "1ofN", "2ofN", etc. By default, keeps
only the "...1of" variant. Also, SLC35E2 is present twice, but without a
numbered annotation. Keeps only the the larger (encompasing) version of
SLC35E2. Set false to keep all versions of all genes (gene name will not
then be a unique key.). Also implies skipUnknownGene}

\item{skipUnknownGene}{Drops unknown genes, those whose names
in the gaf are "?". There are 32 in the version of the GAF this targets.
Set false to keep the unknown genes. Note, this setting is over-ridden to
TRUE (skipping these genes) when uniqueGene= TRUE as all these genes have
the same gene name, "?".}
}
\value{
The main output is the gaf extract file, which is just the
gene models and by default just the genes with unique names. See
\href{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/other/GAF/GAF_bundle_Feb2011/docs/GAF_v2_file_description.docx}{GAF_v2_file_description.docx}
However, this function also returns a list with some summary info about the
gaf and the generated geneModels file:

    info$gaf - The gaf file name used as a parameter (relative file name)
    info$gaf_real - The absolute full path filename to the input gaf file
    info$gaf_md5 - The md5 checksum of the gaf file, as a string
    info$uniqueGene - The uniqueGene parameter setting used.
    info$skipUnknownGene - The skipUnknownGene parameter setting used.
    info$gaf_lines - The number of lines in the gaf file
    info$gaf_models - The number of models in the gaf file
    info$gaf_models_unique - The number of unique models in the gaf file
    info$gaf_extract - The geneModels file name, based on the input gaf
    info$gaf_extract_real - The absolute full path filename of the geneModels file
    info$gaf_extract_md5 - The md5 checksum of the geneModels file, as a # string
}
\description{
This extracts a tab delimited file of all canonical
exon gene models from the GAF. The cannonical model is intended to
be at least a common starting point for a simplified definition of "the gene
called X", based on the union of all transcript. By default each model in the
output file (row) can be refereed to uniquely by gene name.
This requires the "grep" program, so only works on linux/mac systems. The Gaf
file version this works with is the version used for the TCGA RNAseq
expression data files. It is available for download at the NCI uncompressed:
\href{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/other/GAF/GAF.hg19.June2011.bundle/outputs/TCGA.hg19.June2011.gaf}{TCGA.hg19.June2011.gaf}
or gzipped:
\href{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/other/GAF/GAF.hg19.June2011.bundle/outputs/TCGA.hg19.June2011.gaf.gz}{TCGA.hg19.June2011.gaf.gz}
}

