% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gafGeneModels.R
\name{extractTranscriptModels}
\alias{extractTranscriptModels}
\title{Extract transcript models from the gaf}
\usage{
extractTranscriptModels(gafFile, outFile = paste0(gafFile,
  ".transcriptModels"), force = FALSE)
}
\arguments{
\item{gafFile}{The full-path name of the GAF file [REQ]}

\item{outFile}{The output file name, by default the same as the gafFile, with
".geneModels" appended. Will not overwrite unless force= is set true}

\item{force}{If the output file exists, overwrite (generates a warning).}
}
\value{
The number of transcripts in the output file (should be 73,707)
}
\description{
This extracts a tab delimited file of all canonical transcript models from the
GAF. Each model in the output file (row) can be refereed to
uniquely by transcript name.
}

