% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortFusions.R
\name{filterFusions}
\alias{filterFusions}
\title{Sub-select fusion lines from a list of fusions}
\usage{
filterFusions(fusions, sample = c(), gene1 = c(), gene2 = c(),
  genePairing = "or")
}
\arguments{
\item{fusions}{The data frame of fusions to filter}

\item{sample}{List of sample names. If given, only fusions for the
samples with exactly mathcing sample names will be kept. Gene filtering
will select only from these samples.}

\item{gene1}{List of gene names to filter fusions based on the
upstream gene name.

 @param gene2 List of gene names to filter fusions based on the
downstream gene name.

 @param genePairing Logic to use if both gene1 and gene2 given.
   \tabular{ll}{
      \code{"or"} \tab If fusion gene1 in gene1 param or fusion gene2 in
                       gene2 param, keep the fusion.\cr
      \code{"and"} \tab If fusion gene1 in gene1 param and fusion gene2 in
                         gene2 param, keep the fusion.\cr
      \code{"pair"} \tab Keep fusion if param gene1[i] = fusion gene1 and
                           param gene2[i] = fusion gene2, for any gene
                           param i (gene1 and gene2 must be same length)\cr
   }}
}
\value{
The selected rows from the fusion data frame
}
\description{
Filtering is based on sample name or gene. If you need more complex filtering,
just do it yourself. This is a convienience function.
}

