% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\docType{package}
\name{FusionExpressionPlot}
\alias{FusionExpressionPlot}
\alias{FusionExpressionPlot-package}
\title{Plot Expression Profiles for Fused Genes}
\description{
Plots exons from two genes as boxes scaled by exon size and colored based on
expression in a sample. Fusion points between the two genes are represented
as triangles at the appropriate position. This allows visualizing the drop in
relative expression of the exons that are part of a putative expressed fusion
transcript relative to the exons only in the unfused gene transcripts.
Normalization is performed across an entire cohort of samples and then again
within each gene, so color differences show exons whose expression is
unusually variant within the genes against a background of similar samples.
}

