% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohortFusions.R
\name{getCandidateFusions}
\alias{getCandidateFusions}
\title{Select candidate fusions by sample and gene}
\usage{
getCandidateFusions(fusionDF, sample, gene1, gene2)
}
\arguments{
\item{fusionDF}{The data frame of fusion data}

\item{sample}{The vector of sample names to filter by}

\item{gene1}{The vector of gene names to filter by (either end)}

\item{gene2}{Another vector of gene name to filter by (either end)}
}
\value{
The rows from the data frame matching any given sample and where
both fusion gene ends are in the one of the two lists given (in any
combination)
}
\description{
Select out the candidate fusions for the given samples as a new data.
For each samples, include any fusion involving either gene. May not care
about a specific sample-gene1-gene2 combo, but that is for something else
to filter.
}

