% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortNormalizedExpression.R
\name{getCohortExonExpressionData}
\alias{getCohortExonExpressionData}
\title{Generate a data frame of cohort exon expression data}
\usage{
getCohortExonExpressionData(geneModels, cohortFiles, progress = TRUE,
  type = "rpkm")
}
\arguments{
\item{geneModels}{A data frame describing the exon structure of the genes
(gene models). See \code{\link{loadGafModels}}. This must correspond to the
exon expression file exon names.}

\item{cohortFiles}{A data frame describing the samples and their exon
expression files. See \code{\link{loadCohortDefinition}}. All files must
exist, a simple check is \code{all( file.exists(
\var{cohortFiles}$exonExpressionFile))}.}

\item{progress}{By default shows a progress bar. Set this to \code{FALSE} if
don't want this (e.g. if you are logging output when running this in a
batch job).}

\item{type}{The type of expression being used. This is the name of the column
from the exon expression files to use. By default uses the column named
"rpkm". See \code{\link{loadExonExpressionFile}}}
}
\value{
Returns a data frame that concatenates the columns from the gene
models data frame with one column from each sample in the cohort data frame.
}
\description{
Load all the exon expression files for a cohort of samples and build a single
data frame with exon information from a geneModels data frame and one column
of expression information per sample in the cohort. This takes a while, so
progress by sample is reported by default.
}

