% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{getFusionExpressionPlotData}
\alias{getFusionExpressionPlotData}
\title{Generate the plot data object for a fusion expression plot.}
\usage{
getFusionExpressionPlotData(genes, fusions, geneNames, geneSep = 500,
  fusionSep = 40, nameSep = 20)
}
\arguments{
\item{genes}{List of genomic ranges objects to plot, by gene name.}

\item{fusions}{List of vectors of genomic fusion points to plot, by gene
  name.}

\item{geneNames}{Vector of gene names.}

\item{geneSep}{Distance between genes in the plot, in bases. Default = 500.}

\item{fusionSep}{Distance between exon tops and tops of plotted fusion
  points, in bases. Default = 40}

\item{nameSep}{Distance between the plot and the gene name annotation, in
  bases. Default = 20.}
}
\value{
A "fusion expression plot data" object, a list with the data needed
  to plot the gene models, expression levels, and fusion points. This list
  has the folowing elements:

\tabular{ll}{
   geneRects   \tab The coordinates of the rectangles to plot as exons\cr
   fusionPolys \tab The coordinates of the polygons (triangles) used as fusion point markers\cr
   geneNames   \tab The names of the genes being plotted\cr
   xRange      \tab The min and max x values in the plot\cr
   yRange      \tab The min and max y values in the plot\cr
}
}
\description{
Merge data from multiple genes and multiple point sets into a list for
plotting as a fusion expression graph. Uses \code{\link{grToRect}} and
\code{\link{pointToPoly}}. This is a lower-level function called by
plotting functions like \code{\link{plotFusionExpressionOne}} or
\code{\link{plotFusionExpressionPair}}
}
\seealso{
\code{\link{grToRect}}, \code{\link{pointToPoly}},
\code{\link{plotFusionExpressionOne}}, \code{\link{plotFusionExpressionPair}}
\code{\link{plotFusionExpressionPlotData}}
}

