% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohortFusions.R
\name{getMapSpliceCohortFusionData}
\alias{getMapSpliceCohortFusionData}
\title{Load a mapplice cohort fusion file.}
\usage{
getMapSpliceCohortFusionData(file, all.columns = FALSE)
}
\arguments{
\item{file}{The path to the data file to load.}

\item{all.columns}{Set TRUE to load all columns from data file. By
default (all.columns=FALSE) only loads and returns some columns (see below).}
}
\value{
By default only returns the following columns:

\tabular{lll}{
   \code{id} \tab was the row number \tab original row number\cr
   \code{gene1} \tab was \code{D_gene} \tab donor side gene\cr
   \code{gene2} \tab was \code{A_gene} \tab acceptor side gene\cr
   \code{gene1pos} \tab was \code{D_end} \tab Last genomic donor side position\cr
   \code{gene2pos} \tab was \code{A_end} \tab First genomic acceptor side position\cr
   \code{sample} \tab was \code{tcga_id} \tab The sample name\cr
}

If all.columns= TRUE, all of the other columns from the fusion data file
will be kept, with the original names.
}
\description{
Load a mapplice cohort fusion file.
}

