% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{grAddColumn}
\alias{grAddColumn}
\title{Add a data column to a GRanges object.}
\usage{
grAddColumn(gr, name, vec)
}
\arguments{
\item{gr}{The granges object to add to add a column to.}

\item{name}{The name of the column when added}

\item{vec}{The data. If not the same length as the GRanges object, will
attempt to wrap values. If not an even multiple, will error.}
}
\value{
The granges object, with the appended column.
}
\description{
Adds a vector as a column in a GRanges object, assuming the data is in the
correct order. This data can be anything describing the range elements, for
instance a vector of colors. This is essentially a convienience wrapper
around the elementMetadata() function from the GRanges package.
}

