% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{grFromCohortDf}
\alias{grFromCohortDf}
\title{Extract GRange objects from cohort exon expression data}
\usage{
grFromCohortDf(df, genes, sample, column = "exonData")
}
\arguments{
\item{df}{A dataframe with gene model and exon expression. Required to have
  the following columns: 'chr', 'start', 'end', 'strand', 'gene', and one
  column for each sample.}

\item{genes}{The names of the genes to extract from the cohort data frame.}

\item{sample}{The name of the sample to extract from the df. Required}

\item{column}{The metadata column name under which the expression data is
  added to the GRange objects. By default this is 'exonData'}
}
\value{
A List of GRanges object for the gene, with start and end positions
  of the exon specified, and with a metadata column "relativeExpression"
  giving the relative expression for the selected gene in the selected
  sample.
}
\description{
Given a dataframe describing exon expression for a sample cohort, generate a
list of GRange objects for one sample, by selected genes. Each GRange object
is identified by gene name and includes a meta-data column giving the
expression data.
}
\section{Cohort exon expression data}{


  The input dataframe must have the following columns:

  \describe{
      \item{chr}{
  The chromosome this exon is on, as 'chr1'...'chr22', chrX', 'chrY', or
  'chrM'.}
      \item{start}{
  The position of the first base of the exon.}
      \item{end}{
  The position of the last base of the exon.}
      \item{strand}{
  The strand of this exon (+,-, or *). All exons in a gene should have the
  same strand.}
      \item{gene}{
  The name of this gene. All exons in a gene should have the same name, and
  no exons from different genes should have the same name.}
      \item{***}{
  The normalized expression value for this exon in sample named "***". Should
  have one column for every sample in the cohort. Sample names may not be
  "chr", "start", "end", "strand", "gene", and must be unique.}
  }

  Start and end are assumed to be 1 based, but this does not really matter as
  long as all positions here and in any position data (i.e. fusions) used
  with this data all have the same base.
}

