% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grUtils.R
\name{grFromElementString}
\alias{grFromElementString}
\title{Convert a genomic elements string into an GRanges object.}
\usage{
grFromElementString(elementString, chr, strand = "*",
  seqDelim = "\\\\s*,\\\\s*", posDelim = "\\\\s*-\\\\s*")
}
\arguments{
\item{elementString}{The string to parse into genomic ranges.}

\item{chr}{The chromosome for the genomic ranges.}

\item{strand}{The strand for the chromosome, must be '*', '+' or '-'.
Defaults to '*' if not specified.}

\item{seqDelim}{The sequence delimiter, usually use the default '\\s*,\\s*'. Can be
any reqular expression.}

\item{posDelim}{The start-end position separator, usually use the default
  '\\s*-\\s*'. Can be any regular expression.}
}
\value{
A GRanges object based on the input elements.
}
\description{
Parses a string into a sequence of genomic elements, adding specified
chromosome and strand. Each genomic elements is assumed to be two genomic
positions separated by a delimiter (-), with the low position first and the
high position second, in 1 based coordinates. The sequence of elements are
separated by a delimiter (,). All elements must be on the same sequence block
(chromosome) and must be on the same strand (by default '*' for unspecified).
Positions are given as if on the positive strand, actual positions would be
reversed if on the negative strand. Whitespace around delimiters is ignored.
If alternate delimiters that are whitespace are used, then they must be
present but additional whitespace is ignored.
}

