% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grUtils.R
\name{grFromLocationString}
\alias{grFromLocationString}
\title{Convert a genomic location string into a GRange object.}
\usage{
grFromLocationString(locationString,
  captureRE = "(?<chr>.*?)\\\\s*:\\\\s*(?<elements>.*?)\\\\s*:\\\\s*(?<strand>[-+*]?)",
  ...)
}
\arguments{
\item{locationString}{The string to parse into a GRange object}

\item{captureRE}{The regular expression used to extract the substrings for
the chromosome (chr), the element list (elements), and strand (strand).}

\item{...}{Allows passing alternate fromDelim and betweenDelim regular
expressions through to grFromElementString, which parses the element part of
the locationstring}
}
\value{
A GRange object corresponding to the location string.
}
\description{
Given a genomic location string, converts it to a GRanges object. The
location string is parsed based on a regular expression with up to three
capture groups, capturing the chromosome, the strand, and the element
location sub-strings. The element string is then parsed into a sequence of
start and stop ranges, and the this data is used to create and return a
GRange object. The regexp used to match with must contain at least three
named capture groups: \code{(?<chr>..)}, \code{(?<elements>...)}, and
\code{(?<strand>..)}. Matching is done using \code{perl=TRUE}).
}
\details{
The default location string parsed is assumed to look something like:
\preformatted{    "chr1:1-100,200-300,400-500:-"}

White-space is ignored around the delimiters ':', ',', and '-'.
}

