% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grUtils.R
\name{grNew}
\alias{grNew}
\title{Create a simple GRanges object.}
\usage{
grNew(start, end, chr, strand = "*")
}
\arguments{
\item{start}{A vector giving the starting (lower) genomic coordinates of the
exons. Must be specified.}

\item{end}{A vector giving the ending (higher) genomic coordinates of the
exons. Must be specified.}

\item{chr}{The chromosome (singular) on which these exons occur, like 'chr1'
or 'chrX'. Must be specified.}

\item{strand}{The strand the exons are actually on, as '+', '-', or '*'.
Defaults to '*' if not specified.}
}
\value{
The genomicRanges object built from the exons.
}
\description{
Create a simple Granges object based on a sequence of exons on one strand of
one chromosome. This could be a gene, a transcript, or one side of a fusion.
Defaults to '*' if strand is not specified. Coordinates are given as if the
exons were on the positive strand, in 1-based coordinates. This will be
backawards from the actual sequence if the gene is on the negative strand.
Ranges will be sorted by start and then by end.
}

