% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{grToRect}
\alias{grToRect}
\title{Plotable rectangles from GRanges Object}
\usage{
grToRect(gr, intronWidth = 100, exonHeights = 100,
  exonFillColors = "blue", exonFillColorsCol = "exonFillColors",
  exonBorderColors = "black", exonBorderColorsCol = "exonBorderColors",
  exonHeightsCol = "exonHeights")
}
\arguments{
\item{gr}{The GenomicRanges object to plot.

  Object giving the gene to calculate the parameters for. Passed through to
  genomicToModelCoordinates as its gene parameter. Besides the criteria
  specified there, may also contain data columns 'exonHeight',
  'exonFillColor' and 'exonBorderColor'. If present, the values in these
  columns will be used instead on the equivalently names parameters to this
  function.}

\item{intronWidth}{The fixed spacing to use for gaps between rectangles.

  The width of the itron, the x spacing between the exon rectangles. Passed
  through to genomicToModelCoordinates as its gapWidth parameter. Defaults to
  100.}

\item{exonHeights}{The vector of heights to use for the rectangles.

  If fewer than the number of elements are givien, will be wrapped. If not
  specified exonHeightsCol must contain the column name in the GRanges object
  to use instead, by default it looks for column 'exonHeight'.}

\item{exonFillColors}{The vector of fill colors to use for the rectangles.

  If fewer than the number of elements are givien, will be wrapped. If not
  specified exonFillColors must contain the column name in the GRanges object
  to use instead, by default it looks for column 'exonFillColor'.}

\item{exonFillColorsCol}{The name of GRanges column to use for
  exonFillColors, if not specified. By default 'exonFillColor'}

\item{exonBorderColors}{The vector of border colors to use for the
  rectangles.

  If fewer than the number of elements are givien, will be wrapped. If not
  specified exonBorderColors must contain the column name in the GRanges
  object to use instead, by default it looks for column 'exonBorderColor'.}

\item{exonBorderColorsCol}{The name of GRanges column to use for
  exonBorderColors, if not specified. By default 'exonBorderColor'}

\item{exonHeightsCol}{The name of GRanges column to use for exonHeights, if
  not specified. By default 'exonHeight'}
}
\value{
A list with the following elements:

  \describe{
      \item{xStarts}{The low x coordinate; the left side of each rectangle.}
      \item{xEnds}{The high x coordinate; the right side of each rectangle.}
      \item{yBottoms}{The low y coordinate; the bottom of each rectangle.}
      \item{yTops}{The high y coordinate; the top of each rectangle.}
      \item{fillColors}{The fill color for each rectangle.}
      \item{fillColors}{The border color for each rectangle.}
      \item{xRange}{c(<min x value>, <max x value).}
      \item{yRange}{c(<min y value>, <max y value).}
  }
}
\description{
Given a genomicRanges Object, return the data needed to plot it as a sequence
of rectangles based on widths. Gaps are all fixed at a single intronWidth.
Height, fill color, and border color for the exon rectangles can be specified
on a per-exon basis (vector-wrapped), or given in the genomicRanges object
(as metadata colums, by  default exonHeight, exonFillColor, and
exonBorderColor). Returns the rectangle as a list of data, as specified
below. Uses the genomicToModelCoordinates function, passing throught the
parameters needed
}

