% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohortNormalizedExpression.R
\name{loadExonExpressionFile}
\alias{loadExonExpressionFile}
\title{Load a TCGA exon expression data file column as a data frame}
\usage{
loadExonExpressionFile(path, type = "rpkm")
}
\arguments{
\item{path}{The path to the exon expression data file to load}

\item{type}{The name of the column containing the expression data to be used,
  by default this is \code{"rpkm"}. Allowed values are \code{"rpkm"},
  \code{"count"}, and \code{"coverage"}.}
}
\value{
A data frame with the following columns from the exon expression
  file:

\tabular{ll}{
   \code{chr}    \tab The exon's chromosome\cr
   \code{start}  \tab The genomic start coordinate for the exon\cr
   \code{end}    \tab The genomic end coordinate for the exon\cr
   \code{strand} \tab The strand the chromosome is on, one of \code{+ | - | *}\cr
   \code{rpkm | count | coverage} \tab One of the exon expression level columns\cr
}
}
\description{
Loads an exon expression data file as generated for the TCGA into a data
frame. Loads the exon identification information but only one of the
expression columns, by default the \code{"rpkm"} column.
}
\section{Errors}{


The following fatal errors can occur:

\describe{
   \item{
      \command{No such file: \var{path}}
   }{
      The exon expression file named can not be found. Case, spelling
      permission and wrong working dir for relative paths are all common
      errors.
   }
   \item{
      \command{Only type values allowed are: rpkm, count, coverage}
   }{
      You have to pick one of the expression columns from the file.
   }
}
}

