% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gafGeneModels.R
\name{loadGeneModels}
\alias{loadGeneModels}
\title{Load the gene models file into a data frame.}
\usage{
loadGeneModels(file)
}
\arguments{
\item{file}{The name of the gene models file extracted from the GAF}
}
\value{
Data frame with the data from the GAF gene models:

  \tabular{ll}{
      \code{$gene} \tab Feature name\cr
      \code{$chr}  \tab chromosome as chr#, chr##, chrX, chrY, chrM, or chrM_rCRS,
                        or possibly a contig name, like 'GL000228.1'.\cr
      \code{$strand} \tab strand as *, +, or -\cr
      \code{$gstart} \tab first base of gene\cr
      \code{$gend} \tab last base of gene\cr
      \code{$exon} \tab Exon number\cr
      \code{$start} \tab first base of exon\cr
      \code{$end} \tab Last base of exon\cr
      \code{$length} \tab length of the exon in bases\cr
   }
}
\description{
Load the gene models file into a data frame.
}

