% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortNormalizedExpression.R
\name{normExpressionData}
\alias{normExpressionData}
\title{Normalize a cohort of exon expression data}
\usage{
normExpressionData(exonExpressionData)
}
\arguments{
\item{exonExpressionData}{A data frame of exon expression data, with one
column per sample in the cohort. See
\code{\link{getCohortExonExpressionData}}.}
}
\value{
A data frame of exon expression data, but with the data normalized
  (scaled to a standard normal) first within one sample by gene, then across
  all samples by exon.
}
\description{
Normalize a cohort of exon expression data
}

