% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{plotFusionExpressionOne}
\alias{plotFusionExpressionOne}
\title{Plot a one-gene fusion expression plot.}
\usage{
plotFusionExpressionOne(geneName1, fusions1, sample, cohortExpressionDF)
}
\arguments{
\item{geneName1}{The name of the gene being plotted.}

\item{fusions1}{A vector of genomic fusion ends within the \code{geneName1}}

\item{sample}{The sample name, for the plot and for lookup of expression in
the \code{cohortExpressionDF}.}

\item{cohortExpressionDF}{Provides the gene models for the specified gene and
the normalized expression for the specified sample. See
\code{\link{normExpressionData}} for format.}
}
\value{
Nothing
}
\description{
This generates one fusion expression plot file showing the expression of one
gene and associated (possibly multiple, possibly none) fusion points to
non-gene regions For plotting two genes, see
\code{\link{plotFusionExpressionPair}}. Essentially a wrapper around two
low-level functions: \code{\link{getFusionExpressionPlotData}} which
generates the plotting data given expression and fusion information, and
\code{\link{plotFusionExpressionPlotData}} which handles the actual plot
file creation.
}
\seealso{
\code{\link{normExpressionData}},
  \code{\link{plotFusionExpressionPlotData}},
  \code{\link{getFusionExpressionPlotData}},
  \code{\link{getCohortExonExpressionData}}
  \code{\link{plotFusionExpressionPair}}
}

