% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{plotFusionExpressionPair}
\alias{plotFusionExpressionPair}
\title{Plot a two-gene fusion expression plot}
\usage{
plotFusionExpressionPair(geneName1, geneName2, fusions1, fusions2, sample,
  cohortExpressionDF)
}
\description{
Plot a two-gene fusion expression plot
}

