% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{plotFusionExpressionPlotData}
\alias{plotFusionExpressionPlotData}
\title{Plot a fusionExpressionPlotData object}
\usage{
plotFusionExpressionPlotData(data, title, file = "temp.pdf", scale = 1000,
  ratio = 4, scaleOffset = 15)
}
\arguments{
\item{data}{A data object generated by
\code{\link{getFusionExpressionPlotData}}.}

\item{title}{The title to put on the graph.}

\item{file}{The filename (pdf) to save the plot as. By default this is
\code{temp.pdf}}

\item{scale}{The number of "bases" per inch in the plot, default = 1000}

\item{ratio}{The height/width ratio of the plot.}

\item{scaleOffset}{The distance in bases that the scale annotation is set
below the gene exon graph.}
}
\value{
nothing
}
\description{
Plots the data object generated by \code{\link{getFusionExpressionPlotData}}
and saves it as a pdf file. This is a lower-level function called by plotting
functions like \code{\link{plotFusionExpressionOne}} or
\code{\link{plotFusionExpressionPair}}
}
\seealso{
\code{\link{getFusionExpressionPlotData}}
  \code{\link{plotFusionExpressionOne}}
  \code{\link{plotFusionExpressionPair}}
}

