% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohortFusions.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Plot the expression of a gene}
\usage{
plotGeneExpression(samples, genes, cohortExonExpression)
}
\arguments{
\item{samples}{Vector of sample names giving the expression column from the
  cohortExonExpression to plot. Must be the same length as the gene parameter}

\item{genes}{Vector of gene names giving the exon rows from the
  cohortExonExpression to plot. Must be the same length as the sample
  parameter}

\item{cohortExonExpression}{= cohort expression data frame, with the first 9
  columns giving the gene model with one exon per row, and the remaining
  columns giving exon expression for different samples with their headings
  the sample names.}
}
\value{
Nothing
}
\description{
For each pair of sample and gene, plot the expression for that gene in that
sample.
}

