% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{pointToPoly}
\alias{pointToPoly}
\title{Convert points to polygons}
\usage{
pointToPoly(x, y, widths = 20, heights = 20, shapes = "triangleDown",
  sides = 5, fillColors = "red", borderColors = "black", scale = c(4,
  1), resScale = c(1, 1))
}
\arguments{
\item{x}{The x plot coordinates where the the poly-points should be drawn}

\item{y}{The y plot coordinates where the the poly-points should be drawn}

\item{widths}{The widths in plot coordinates of the poly-points. Default =
5.}

\item{heights}{The height in plot coordinates of the poly-points. Default =
5.}

\item{shapes}{The shape of the poly-point. Allowed shapes are
\code{'triangleDown'} (the default) and \code{'triangleUp'}.}

\item{sides}{The position of the poly-points over the coordinate point. Can
have a value from 1-9, oriented like a 3 x 3 touch pad, 1 has uper left
corner of poly as the point, 5 has the middle over the point, and 9 has the
lower right as the point. Default = 5, centered.}

\item{fillColors}{The colors filling the poly-point. Default = \code{'red'}.}

\item{borderColors}{The colors of the poly-point borders. Default =
\code{'black'}.}

\item{scale}{The scale factor of the x and y axis of the plot, i.e if plot
is 100x500, scale should be \code{c(1,5)} or \code{c(0.2,1)}. By default
this is \code{c(1,1)}.}

\item{resScale}{The scale factor of the device, i.e. if the screen
resolution is 1024x768, this should be \code{c(1,1024/768)} or
\code{c(768/1024,1)}. By default this is \code{c(1,1)}.}
}
\value{
Returns a list with the data needed to plot the points as polygons.
  The elements of the list are:

\tabular{ll}{
   \code{count} \tab number of points plotted as polygons\cr
   \code{xRadii} \tab 1/2 widths of points, after wrapping\cr
   \code{yRadii} \tab 1/2 heights of points, after wrapping\cr
   \code{shapes} \tab The shapes parameter, after wrapping\cr
   \code{sides} \tab The sides parameter, after wrapping\cr
   \code{fillColors} \tab The fillColors parameter, after wrapping\cr
   \code{borderColors} \tab The borderColors parameter, after wrapping\cr
   \code{x} \tab X coordinates of polygons to draw, \code{NA} separated\cr
   \code{y} \tab Y coordinates of polygons to draw, \code{NA} separated\cr
   \code{xRange} \tab lowest and highest x value drawn\cr
   \code{yRange} \tab lowest and highest y values drawn\cr
}
}
\description{
Converts a vector of coordinate points to a data object that can be drawn
with \code{\link{polygon}}. It calculate the values need to plot the points
as polygons specified by shape. Currently only triangles pointing up and
triangles pointing down can be specified. Each point can be described
separately by passing in a vector for the \code{shapes} parameter. Other than
x and y coordinates of the points to plot and the two scale paramters,
provided values will be wrapped as needed.
}
\section{Shape scaling}{


  The shape of a pixel is not necessarily square, so a square of 10 x 10
  pixels may not be square but instead rectangular. To make the polygon
  shapes come out symetrical, the relative heights and widths of the pixels
  must be specified. This is the purpose of the \code{scale} and
  \code{resScale} parameters. Two parameters are provided to allow specifying
  the plot scaling ratio and the screen resolution scaling ratio separately.
}

