% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{sortDataIntoBins}
\alias{sortDataIntoBins}
\title{Sort values into bins}
\usage{
sortDataIntoBins(data, binEnds, binLabels)
}
\arguments{
\item{data}{A vector of numeric data to bin.}

\item{binEnds}{The n+1 ends of the n bins into which the data is sorted.}

\item{binLabels}{The vector of values corresponding to the bins to}
}
\value{
The vector of bin labels corresponding to the bins the data was
sorted into.
}
\description{
Converts a vector of numeric values into a vector of bin names, where the
n bins to sort into are defined by n+1 bin boundaries (including start and
end boudaries, probably -Inf and Inf). Good for assigning a gradient of
colors to values.
}
\details{
Bins the data values into n bins as s specified by the n+1 binEnds, then
returns a vector of the same length as data, but with each value replaced by
the binLabel for the bin it was sorted into. E.g if the
bin labels are colors, this converts data into a color scale, or if the bin
labels are the upper (or lower bin ends), this converts values into their
}

