% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohortFusions.R
\name{do.allPlots}
\alias{do.allPlots}
\title{Plot a fusion fusion expression plot for every fusion}
\usage{
do.allPlots(fusions, normalizedCohortExpression)
}
\arguments{
\item{fusions}{The fusions to plot}

\item{normalizedCohortExpression}{The normalized exon expression cohort data
  frame.}
}
\value{
A list with two entries

\tabular{ll}{
   \code{plotCount} \tab The number of plots generated.\cr
   \code{fusionLineCount} \tab The number of fusions in the list to plot.\cr
}
}
\description{
Plot a fusion fusion expression plot for every fusion
}

