% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grUtils.R
\name{grFromElementString}
\alias{grFromElementString}
\title{Convert a genomic element string into an GRanges object.}
\usage{
grFromElementString(elementString, chr, strand = "*",
  betweenDelim = "\\\\s*,\\\\s*", fromDelim = "\\\\s*-\\\\s*")
}
\arguments{
\item{elementString}{The string to parse into genomic ranges.}

\item{chr}{The chromosome for the genomic ranges.}

\item{strand}{The strand for the chromosome, must be '*', '+' or '-'.
Defaults to '*' if not specified.}

\item{betweenDelim}{The sequence delimiter, usually use the default '\\s*,\\s*'. Can be
any regular expression.}

\item{fromDelim}{The start-end position separator, usually use the default
  '\\s*-\\s*'. Can be any regular expression.}
}
\value{
A GRanges object based on the input elements, with ranges sorted by start and then
end position
}
\description{
Parses a genomic elements string into ranges, then adds a specified
chromosome and strand and returns a GRanges object. The elements string parsed
by default looks like:
\preformatted{    "1-100,3-500,2-3" }
}
\details{
Spaces around the between delimiter ',' and the from delimiter '-' are ignored;
other delimiters (regular expressions) can be specified.  All elements must be on the same sequence block
(chromosome) and must be on the same strand (by default '*' for unspecified).
Positions are given in 1 based coordinates as if on the positive strand with the low position first and the
high position second. Actual positions would be
reversed if on the negative strand.
}

