% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohortNormalizedExpression.R
\name{loadCohortDefinition}
\alias{loadCohortDefinition}
\title{Read in the cohort definition file}
\usage{
loadCohortDefinition(file, samples = NULL, comment.char = "#",
  dataRoot = NULL)
}
\arguments{
\item{file}{The name of the cohort file to read in (tab-delimited with
  header)}

\item{samples}{The list of samples to read in, as a character vector. The rows
  matching these are the only ones read in. By default, this is set to
  \code{NULL}, meaning it reads all the samples. If any sample in the
  provided list is not found, a warning will be generated. Must match exactly
  (case sensitive).}

\item{comment.char}{The character starting comment lines, by default
  \code{'#'}. Must be a single character. Any line beginning with this
  character is considered to be a comment line and is ignored. Must be the
  first character in a line (no leading white space is allowed). Set to the
  empty text, \code{""}, to skip comment line filtering.}

\item{dataRoot}{A directory to prepended to the \code{exonExpressionFile}
  filenames. By default this is \code{NULL} and nothing is prepended.
  \code{NA} and empty text \code{""} are liekwise ignored. It is ok if
  \code{dataRoot} does not exist at this time as the file locations are not
  verified at this point.}
}
\value{
A data frame representing the sample cohort with at least two
  columns:

\tabular{ll}{
   \code{sample}
      \tab The sample names read from the input file's \code{sample} column\cr
  \code{exonExpressionFile}
      \tab The exon expression file names read from the input file's
            \code{exonExpressionFile} column\cr
}
}
\description{
Creates a data frame from a tab-delimited file with a header describing a
sample cohort. The cohort file must specify at minimum columns named 'sample'
and 'exonExpressionFile'. These names must match case sensitively.
Allows selecting a subset of the samples to keep. Will skip comment
lines. All columns from the cohort file are loaded, but only the sample and
exonExpressionFile columns are validated.
}
\section{Errors}{


\describe{
   \item{
      \command{Can't find the cohort definition file: \var{file}}
   }{
      The file you are trying to read in can't be found - case, permission,
      relative directory, and misspellings are all possible reasons.
   }
   \item{
      \command{The cohort file has no sample column \var{sample}}
   }{
      The cohort file must have a column \code{sample} with the sample names
   }
   \item{
      \command{ The cohort file has no expression file column \var{exonExpressionFile}}
   }{
      The cohort file must have a column \code{exonExpressionFile} giving the file names
      of the exon expression data for the sample in the same row.
   }
   \item{
      \command{The sample and exon expression file columns must contain text}
   }{
      Either or both of these columns was read in as something other than
      a character vector.
   }
   \item{
      \command{The cohort file can not contains duplicate samples}
   }{
      The cohort file contains multiple lines with the same sample name, even
      after filtering out any samples you didn't want considered. That's not
      ok.
   }
   \item{
      \command{The cohort file can not contain duplicate cohort expression
         files}
   }{
      The cohort file contains multiple lines with the same cohort expression
      file name, even after filtering out any samples you didn't want
      considered. That's not ok. If two different samples really have the
      same expression file name, you'll have to put them in different
      directories.
   }
   \item{
      \command{All sample and exon expression file entries must be non-empty
         text}
   }{
      Can not have missing or empty text for samples as this is the
      primary key for later work. Can not have missing or empty text for
      exon expression files (after filtering) as such samples can not then
      be part of the cohort.
   }
}
}

\section{Warnings}{


\describe{
   \item{
      \command{Duplicates in "samples" parameter filter list ignored}
   }{
      The list of samples you provided to filter the cohort data frame by
      contained the same sample name more than once. You can only include
      a sample once, so the duplicates are just ignored. Warning you allows
      this to be changed for future use (if you care).
   }
   \item{
      \command{Ignoring missing samples specified in the "samples" parameter
         list: \var{sample, sample,...}}
   }{
      The list of samples you provided to filter the cohort data frame
      contained sample names that were not actually in the cohort file.
      Perhaps you should double check your cohort file?
   }
}
}

\section{Todo}{

\itemize{
   \item{Consider general validator tool with key column and column
   validation selectable or specifiable. }
}
}

