% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortNormalizedExpression.R
\name{loadExonExpressionFile}
\alias{loadExonExpressionFile}
\title{Load a TCGA exon expression data file as a data frame}
\usage{
loadExonExpressionFile(path)
}
\arguments{
\item{path}{The path to the exon expression data file to load}
}
\value{
A data frame with the following columns from the exon expression
  file:

\tabular{ll}{
   \code{chr}      \tab The exon's chromosome\cr
   \code{start}    \tab The genomic start coordinate for the exon\cr
   \code{end}      \tab The genomic end coordinate for the exon\cr
   \code{strand}   \tab The strand the chromosome is on, one of \code{+ | - | *}\cr
   \code{count}   \tab One of the exon expression level columns\cr
   \code{coverage} \tab One of the exon expression level columns\cr
   \code{rpkm}     \tab One of the exon expression level columns\cr
}
}
\description{
Loads an RNA exon expression data file as generated for the TCGA, returning
its contents as a data frame. The input file has four tab-delimited columns,
no header. The first column in the file describes the exon as
<chr>:<start>-<end><strand>, the next three give count, coverage, and rpkm
values.
}
\section{Errors}{


The following fatal errors can occur:

\describe{
   \item{
      \command{No such file: \var{path}}
   }{
      The exon expression file named can not be found. Case, spelling
      permission and wrong working dir for relative paths are all common
      errors.
   }
}
}

