% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gafGeneModels.R
\name{loadGafModels}
\alias{loadGafModels}
\title{Load a GAF feature file into an exon data frame.}
\usage{
loadGafModels(file)
}
\arguments{
\item{file}{The name of the transcrpt models file extracted from the GAF}
}
\value{
Data frame with the data from the GAF transcrpt models:
/describe{
   {$gene}{gene name}
   {$chr}{chromosome as chr#, chr##, chrX, chrY, chrM, or chrM_rCRS}
   {$strand}{strand as *, +, or -}
   {$gstart}{first base of gene}
   {$gend}{last base of gene}
   {$exon}{Exon number}
   {$start}{first base of exon}
   {$end}{Last base of exon}
   {$length}{length of the exon in bases}
}
}
\description{
Load a GAF feature file into an exon data frame.
}

