% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gafGeneModels.R
\name{loadGeneModels}
\alias{loadGeneModels}
\title{Load the gene models file into a data frame.}
\usage{
loadGeneModels(file)
}
\arguments{
\item{file}{The name of the gene models file extracted from the GAF}
}
\value{
Data frame with the data from the GAF gene models:

      $gene = gene name
      $chr = chromosome as chr#, chr##, chrX, chrY, chrM, or chrM_rCRS
      $strand= strand as *, +, or -
      $gstart=  first base of gene
      $gend=  last base of gene
      $exon= Exon number
      $start= first base of exon
      $end= Last base of exon
      $length= length of the exon in bases
}
\description{
Load the gene models file into a data frame.
}

