% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataRangeReductionUtils.R
\name{mapLabels}
\alias{mapLabels}
\title{Map labels to values}
\usage{
mapLabels(data, binEnds, binLabels)
}
\arguments{
\item{binEnds}{The n+1 ends of the n bins into which the data is sorted. One
  bin would be c(-Inf, Inf). Two bins ~ equally weighted would be c(-Inf,
  median(data), Inf). This vector will be sorted before assigning labels, so
  reverse the label order instead of providing bins as high to low. Bin ends
  belong to the lower bin, with the lowest bin end also part of the lowest
  bin.}

\item{binLabels}{The vector of names corresponding to the labels of the bin a
  value belongs in.}

\item{x}{A vector of numeric data to bin. NA values and values above or
  below the lowest bin end are allowed, they convert to NA's in the returned
  label vector.}
}
\value{
The vector of bin labels corresponding to the bins the data was
  sorted into.
}
\description{
Maps labels to a vector of numeric data based on a
vector of bin ends. Each label corresponds to one of n bins defined by n+1
bin boundaries (including start and end boudaries, probably -Inf and Inf).
Returns a vector of the same length as data, but with each value replaced by
the binLabel for the bin it was sorted into. E.g if the bin labels are
categories like low, medium and high, this converts data to categories. Values equal to bin boundaries
are put in the lower bin, with the lowest bin boundary also included in the
lowest bin.
}
\details{
Basically this is a convienience wrapper around cut().
}

