% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fusionExpressionPlot.R
\name{plotFusionExpressionOne}
\alias{plotFusionExpressionOne}
\title{Plot a one-gene fusion expression plot}
\usage{
plotFusionExpressionOne(geneName1, fusions1, sample, cohortExpressionDF)
}
\description{
Plot a one-gene fusion expression plot
}

